// frontend/src/app/core/api/model/userSettingsResponse.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserSettingsResponse { 
    theme?: string;
    language?: string;
    autoLogoutMinutes?: number;
    readOnlyMode?: boolean;
}