// frontend/src/app/core/api/model/accountDeletionRequest.ts
/**
 * Rev-PasswordManager API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AccountDeletionRequest { 
    masterPassword: string;
    confirmation?: boolean;
}